+++ erdjava/pom.xml
@@ -63,6 +63,12 @@
       <artifactId>okhttp</artifactId>
       <version>3.14.2</version>
     </dependency>
+    <dependency>
+      <groupId>org.junit.jupiter</groupId>
+      <artifactId>junit-jupiter-engine</artifactId>
+      <version>5.4.0</version>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
   <build>
